+++ chromium/gpu/command_buffer/service/command_buffer_service.cc
@@ -22,15 +22,15 @@ namespace {
 
 class MemoryBufferBacking : public BufferBacking {
  public:
-  explicit MemoryBufferBacking(size_t size)
+  explicit MemoryBufferBacking(uint32_t size)
       : memory_(new char[size]), size_(size) {}
   ~MemoryBufferBacking() override = default;
   void* GetMemory() const override { return memory_.get(); }
-  size_t GetSize() const override { return size_; }
+  uint32_t GetSize() const override { return size_; }
 
  private:
   std::unique_ptr<char[]> memory_;
-  size_t size_;
+  uint32_t size_;
   DISALLOW_COPY_AND_ASSIGN(MemoryBufferBacking);
 };
 
@@ -124,7 +124,7 @@ void CommandBufferService::SetGetBuffer(int32_t transfer_buffer_id) {
   // This means ring_buffer_ can be NULL.
   ring_buffer_ = GetTransferBuffer(transfer_buffer_id);
   if (ring_buffer_) {
-    int32_t size = ring_buffer_->size();
+    uint32_t size = ring_buffer_->size();
     volatile void* memory = ring_buffer_->memory();
     // check proper alignments.
     DCHECK_EQ(
